Refactor code in separated files
